# openapi/openapi.v1.0.0-rc1.json
{
  "components": {
    "schemas": {
...
      "AlertsResponse": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Alerts",
            "type": "array"
          },
          "rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "title": "AlertsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MT5AccountResponse": {
        "properties": {
          "balance": {
            "title": "Balance",
            "type": "string"
          },
          "company": {
            "default": "",
            "title": "Company",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "equity": {
            "title": "Equity",
            "type": "string"
          },
          "leverage": {
            "title": "Leverage",
            "type": "integer"
          },
          "login": {
            "title": "Login",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "server": {
            "title": "Server",
            "type": "string"
          },
          "trade_mode": {
            "default": "",
            "title": "Trade Mode",
            "type": "string"
          }
        },
        "required": [
          "login",
          "name",
          "server",
          "currency",
          "leverage",
          "balance",
          "equity"
        ],
        "title": "MT5AccountResponse",
        "type": "object"
      },
      "MT5CandleResponse": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "string"
          },
          "low": {
            "title": "Low",
            "type": "string"
          },
          "open": {
            "title": "Open",
            "type": "string"
          },
          "open_time": {
            "format": "date-time",
            "title": "Open Time",
            "type": "string"
          },
          "spread_points": {
            "default": 0,
            "title": "Spread Points",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "tick_volume": {
            "default": 0,
            "title": "Tick Volume",
            "type": "integer"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "timeframe",
          "open_time",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "MT5CandleResponse",
        "type": "object"
      },
      "MT5ConnectRequest": {
        "properties": {
          "login": {
            "exclusiveMinimum": 0.0,
            "title": "Login",
            "type": "integer"
          },
          "password": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "path": {
            "default": "",
            "maxLength": 512,
            "title": "Path",
            "type": "string"
          },
          "server": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Server",
            "type": "string"
          }
        },
        "required": [
          "login",
          "password",
          "server"
        ],
        "title": "MT5ConnectRequest",
        "type": "object"
      },
      "MT5ConnectionResponse": {
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "History",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "login": {
            "title": "Login",
            "type": "integer"
          },
          "login_status": {
            "title": "Login Status",
            "type": "string"
          },
          "server": {
            "title": "Server",
            "type": "string"
          },
          "session_ref": {
            "default": "",
            "title": "Session Ref",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terminal_build": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Build"
          },
          "terminal_version": {
            "default": "",
            "title": "Terminal Version",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "login",
          "server",
          "status",
          "connected",
          "login_status"
        ],
        "title": "MT5ConnectionResponse",
        "type": "object"
      },
      "MT5OrderCalculateResponse": {
        "properties": {
          "estimated_profit": {
            "title": "Estimated Profit",
            "type": "string"
          },
          "expected_margin": {
            "title": "Expected Margin",
            "type": "string"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "title": "Messages",
            "type": "array"
          },
          "order_type": {
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "request_snapshot": {
            "additionalProperties": true,
            "title": "Request Snapshot",
            "type": "object"
          },
          "retcode": {
            "title": "Retcode",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side",
          "order_type",
          "volume",
          "price",
          "expected_margin",
          "estimated_profit",
          "retcode"
        ],
        "title": "MT5OrderCalculateResponse",
        "type": "object"
      },
      "MT5OrderValidateRequest": {
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 64,
            "title": "Comment",
            "type": "string"
          },
          "magic": {
            "default": 0,
            "minimum": 0.0,
            "title": "Magic",
            "type": "integer"
          },
          "order_type": {
            "default": "market",
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "side": {
            "description": "buy | sell",
            "title": "Side",
            "type": "string"
          },
          "slippage": {
            "default": 10,
            "minimum": 0.0,
            "title": "Slippage",
            "type": "integer"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "volume": {
            "default": "0.01",
            "maxLength": 32,
            "minLength": 1,
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side"
        ],
        "title": "MT5OrderValidateRequest",
        "type": "object"
      },
...
      "MT5StatusResponse": {
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "last_error": {
            "default": "",
            "title": "Last Error",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "login": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Login"
          },
          "login_status": {
            "default": "logged_out",
            "title": "Login Status",
            "type": "string"
          },
          "server": {
            "default": "",
            "title": "Server",
            "type": "string"
          },
          "session_ref": {
            "default": "",
            "title": "Session Ref",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terminal_build": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Build"
          },
          "terminal_version": {
            "default": "",
            "title": "Terminal Version",
            "type": "string"
          }
        },
        "required": [
          "connected",
          "status"
        ],
        "title": "MT5StatusResponse",
        "type": "object"
      },
...
      "MT5TickResponse": {
        "properties": {
          "ask": {
            "title": "Ask",
            "type": "string"
          },
          "bid": {
            "title": "Bid",
            "type": "string"
          },
          "spread": {
            "title": "Spread",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "volume": {
            "default": "0",
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "bid",
          "ask",
          "spread",
          "timestamp"
        ],
        "title": "MT5TickResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QuantForg v1.0.0 Release Candidate 1 OpenAPI specification. EXECUTION_ENABLED remains false. No AI endpoints. Generated from FastAPI route registrations.",
    "title": "QuantForg API",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mt5/account": {
      "get": {
        "operationId": "mt5_account_api_v1_mt5_account_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Account",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/candles/{symbol}": {
      "get": {
        "operationId": "mt5_candles_api_v1_mt5_candles__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "H1",
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_pos",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Pos"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MT5CandleResponse"
                  },
                  "title": "Response Mt5 Candles Api V1 Mt5 Candles  Symbol  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Candles",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/connect": {
      "post": {
        "operationId": "mt5_connect_api_v1_mt5_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MT5ConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5ConnectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Connect",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/disconnect": {
      "post": {
        "operationId": "mt5_disconnect_api_v1_mt5_disconnect_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Disconnect",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/order/calculate": {
      "post": {
        "operationId": "mt5_order_calculate_api_v1_mt5_order_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MT5OrderValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5OrderCalculateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Order Calculate",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/order/validate": {
      "post": {
        "operationId": "mt5_order_validate_api_v1_mt5_order_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MT5OrderValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5OrderValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Order Validate",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/status": {
      "get": {
        "operationId": "mt5_status_api_v1_mt5_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Status",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/symbols": {
      "get": {
        "operationId": "mt5_symbols_api_v1_mt5_symbols_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MT5SymbolResponse"
                  },
                  "title": "Response Mt5 Symbols Api V1 Mt5 Symbols Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Symbols",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/symbols/{symbol}": {
      "get": {
        "operationId": "mt5_symbol_info_api_v1_mt5_symbols__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5SymbolResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Symbol Info",
        "tags": [
          "mt5"
        ]
      }
    },
    "/api/v1/mt5/ticks/{symbol}": {
      "get": {
        "operationId": "mt5_latest_tick_api_v1_mt5_ticks__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MT5TickResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mt5 Latest Tick",
        "tags": [
          "mt5"
        ]
      }
    },
...
    "/api/v1/ops/alerts": {
      "get": {
        "description": "Alert rules and recent alerts (info / warning / critical).",
        "operationId": "ops_alerts_api_v1_ops_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Alerts",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/audit": {
      "get": {
        "description": "Audit Center: auth, broker, strategy, risk, execution, paper events.",
        "operationId": "audit_center_api_v1_ops_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCenterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Center",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}